Add checksum checking to GPS message parser (#15)
It has been observed that the UART messages can get corrupted sometimes
and making their ways to the bus.

Before this commit, the parser sends out a CAN message as soon as a
message field is parsed. With this change all the messages are sent
together at the end of the message after checksum check is passed.

This has an added benefit that all GPS CAN messages from the same NMEA
message would get the same timestamp.